+++ CodexLib/BpCache.cs
@@ -1,14 +1,14 @@
-﻿using System;
+﻿using Kingmaker.Blueprints.Items;
+using Kingmaker.Blueprints.Items.Ecnchantments;
+using Kingmaker.Blueprints.JsonSystem;
+using System;
+using System.Collections;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System.Collections;
-using Kingmaker.Blueprints.Items.Ecnchantments;
-using Kingmaker.Blueprints.Items;
-using System.IO;
-using Kingmaker.Blueprints.JsonSystem;
-using System.Collections.ObjectModel;
 
 namespace CodexLib
 {
@@ -45,7 +45,7 @@ namespace CodexLib
         public static IList Get(Type type)
         {
             if (type == null)
-                return new List<object>();
+                return Array.Empty<object>();
             if (_blueprints.TryGetValue(type, out var list))
                 return (IList)Activator.CreateInstance(typeof(ReadOnlyCollection<>).MakeGenericType(type), list);
             return Get(type.BaseType);
@@ -62,7 +62,7 @@ namespace CodexLib
             IList get(Type type)
             {
                 if (type == null)
-                    return new List<object>();
+                    return Array.Empty<object>();
                 if (_blueprints.TryGetValue(type, out var list))
                     return list;
                 return get(type.BaseType);
@@ -78,6 +78,7 @@ namespace CodexLib
         {
             try
             {
+                var watch = Stopwatch.StartNew();
                 Dictionary<Type, IList> dic = new();
                 foreach (var type in types ?? DefaultTypes)
                     dic.Add(type, (IList)Activator.CreateInstance(typeof(List<>).MakeGenericType(type)));
@@ -105,55 +106,65 @@ namespace CodexLib
                     }
                 }
 
-                using (var writer = new BinaryWriter(new FileStream(path, FileMode.CreateNew, FileAccess.Write), Encoding.ASCII))
+                var buffer = new byte[16];
+                using var writer = new FileStream(path, FileMode.CreateNew, FileAccess.Write);
+                foreach (var (type, list) in dic)
                 {
-                    foreach (var (type, list) in dic)
-                    {
-                        // print length of type; print type
-                        string name = type.GetFullName();
-                        writer.Write(name.Length);
-                        writer.Write(name);
+                    // print length of type; print type
+                    string name = type.GetFullName();
+                    writer.Write(name.Length, buffer);
+                    writer.Write(name, buffer);
 
-                        // print number of guids; print guid
-                        writer.Write(list.Count);
-                        foreach (object obj in list)
-                            writer.Write(((SimpleBlueprint)obj).AssetGuid.ToByteArray());
-                    }
-                    writer.Close();
+                    // print number of guids; print guid
+                    writer.Write(list.Count, buffer);
+                    foreach (object obj in list)
+                        writer.Write(((SimpleBlueprint)obj).AssetGuid.ToByteArray(), 0, 16);
                 }
+                writer.Dispose();
+
+                watch.Stop();
+                Helper.Print($"Finished export in milliseconds: {watch.ElapsedMilliseconds}");
             }
             catch (Exception e) { Helper.PrintException(e); }
         }
 
-        private static void LoadResources()
+        /// <summary>
+        /// Load blueprints from stream. Format must be as ExportResources().<br/>
+        /// Tries to load embedded CodexLib.Resources.Blueprints.bin, if null.
+        /// </summary>
+        public static void LoadResources(Stream reader = null, bool checkDupe = true)
         {
-            if (IsLoaded)
-                return;
             try
             {
+                var watch = Stopwatch.StartNew();
                 var bpcache = ResourcesLibrary.BlueprintsCache;
                 var guid = new byte[16];
                 var buffer = new byte[1024];
                 int length;
 
                 //using (var reader = new FileStream("Blueprints.bin", FileMode.Open, FileAccess.Read))
-                using (var reader = Assembly.GetExecutingAssembly().GetManifestResourceStream("CodexLib.Resources.Blueprints.bin"))
+                reader ??= Assembly.GetExecutingAssembly().GetManifestResourceStream("CodexLib.Resources.Blueprints.bin");
+
+                while(true)
                 {
-                    // read length of type
-                    reader.Read(buffer, 0, 4);
+                    // read length of type, exit if end of stream
+                    length = reader.Read(buffer, 0, 4);
+                    if (length == 0)
+                        break;
                     length = BitConverter.ToInt32(buffer, 0);
                     if (length > 1024)
-                        throw new FormatException("Type length too large");
+                        throw new FormatException($"Type length too large: {length}");
 
                     // read type
                     reader.Read(buffer, 0, length);
-                    var type = Type.GetType(Encoding.ASCII.GetString(buffer, 0, length));
+                    string typeName = Encoding.ASCII.GetString(buffer, 0, length);
+                    var type = Type.GetType(typeName);
                     if (type == null)
-                        throw new FormatException("Type couldn't be parsed");
+                        throw new FormatException($"Type couldn't be parsed '{typeName}', length={length}");
 
                     // get list
                     if (_blueprints.Ensure(type, out var list, typeof(List<>).MakeGenericType(type)))
-                        Helper.Print($"_blueprint missing type: {type.Name}");
+                        Helper.PrintDebug($"Adding type list: {type.Name}");
 
                     // read number of guids
                     reader.Read(buffer, 0, 4);
@@ -177,14 +188,17 @@ namespace CodexLib
                             Helper.PrintDebug($"{bp.AssetGuid} type={bp.GetType().Name} is not {type.Name}");
                             continue;
                         }
-#if DEBUG
+
                         if (list.Contains(bp))
                             Helper.PrintDebug("duplicate blueprint: " + bp.AssetGuid);
                         else
-#endif
                             list.Add(bp);
                     }
                 }
+                reader.Dispose();
+
+                watch.Stop();
+                Helper.Print($"Loaded blueprints in {watch.ElapsedMilliseconds}ms {_blueprints.Join(f => $"{f.Key.Name}:{f.Value.Count}")}");
             }
             catch (Exception e) { Helper.PrintException(e); }
             IsLoaded = true;
@@ -195,34 +209,36 @@ namespace CodexLib
         /// </summary>
         public static void AddResource(SimpleBlueprint bp)
         {
-            if (bp == null)
-                return;
+            try
+            {
+                if (bp == null)
+                    return;
 
-            var type = bp.GetType();
-            foreach (var entry in _blueprints)
-                if (entry.Key.IsAssignableFrom(type) && !entry.Value.Contains(bp))
-                    entry.Value.Add(bp);
+                var type = bp.GetType();
+                foreach (var entry in _blueprints)
+                    if (entry.Key.IsAssignableFrom(type) && !entry.Value.Contains(bp))
+                        entry.Value.Add(bp);
+            }
+            catch (Exception e) { Helper.PrintException(e); }
         }
 
         [HarmonyPatch(typeof(BlueprintsCache), nameof(BlueprintsCache.Init))]
         [HarmonyPostfix]
         private static void OnInit()
         {
-            LoadResources();
+            if (IsLoaded)
+            {
+                Helper.PrintError("BlueprintsCache already initiated!");
+                return;
+            }
+            LoadResources(null, false);
         }
 
         [HarmonyPatch(typeof(BlueprintsCache), nameof(BlueprintsCache.AddCachedBlueprint))]
         [HarmonyPostfix]
         private static void OnAddCachedBlueprint(SimpleBlueprint bp)
         {
-            try
-            {
-                AddResource(bp);
-            }
-            catch (Exception e)
-            {
-                Helper.PrintException(e);
-            }
+            AddResource(bp);
         }
     }
 }