+++ CodexLib/changelog.md
@@ -1,5 +1,8 @@
 ﻿# Changelog
 
+## [1.3.3]
+- added BpCache
+
 ## [1.3.2]
 - compiled for 2.0.4j
 - removed Patch_DuelistParry